Rename models to types
