+++ src/actions/clusterWizard.ts
@@ -1,4 +1,4 @@
 import { createAction } from 'typesafe-actions';
-import { WizardStep } from '../models/wizard';
+import { WizardStep } from '../types/wizard';
 
 export const setCurrentStep = createAction('SET_CURRENT_WIZARD_STEP')<WizardStep>();
