+++ src/api/hosts.ts
@@ -1,7 +1,7 @@
 import axios, { AxiosPromise } from 'axios';
-import { Host } from '../models/hosts';
-import { K8sListApiResponse } from './index';
+import { Host } from '../types/hosts';
+import { ListApiResponse } from './index';
 
-type GetHostsApiResponse = AxiosPromise<K8sListApiResponse<Host>>;
+type GetHostsApiResponse = AxiosPromise<ListApiResponse<Host>>;
 
 export const getHosts = (): GetHostsApiResponse => axios.get('/api/hosts');
